+++ apps/cowswap-frontend/src/modules/account/containers/AccountDetails/index.tsx
@@ -5,7 +5,6 @@ import { getEtherscanLink, getExplorerLabel, shortenAddress, getExplorerAddressL
 import { Command } from '@cowprotocol/types'
 import { ExternalLink } from '@cowprotocol/ui'
 import {
-  ConnectionType,
   useWalletInfo,
   useWalletDetails,
   useIsWalletConnect,
@@ -13,6 +12,7 @@ import {
   useDisconnectWallet,
   useConnectionType,
   getIsInjectedMobileBrowser,
+  ConnectionType,
 } from '@cowprotocol/wallet'
 
 import { Trans } from '@lingui/macro'
@@ -75,7 +75,6 @@ export interface AccountDetailsProps {
   confirmedTransactions: string[]
   ENSName?: string
   forceHardwareWallet?: boolean
-  toggleWalletModal: Command
   toggleAccountSelectorModal: Command
   handleCloseOrdersPanel: Command
 }
@@ -84,7 +83,6 @@ export function AccountDetails({
   pendingTransactions = [],
   confirmedTransactions = [],
   ENSName,
-  toggleWalletModal,
   toggleAccountSelectorModal,
   handleCloseOrdersPanel,
   forceHardwareWallet,
@@ -179,17 +177,10 @@ export function AccountDetails({
                     </AddressLink>
                   )}
 
-                  {standaloneMode !== false && (
-                    <>
-                      {connectionType !== ConnectionType.GNOSIS_SAFE && (
-                        <WalletAction onClick={toggleWalletModal}>
-                          <Trans>Change Wallet</Trans>
-                        </WalletAction>
-                      )}
-                      <WalletAction onClick={handleDisconnectClick}>
-                        <Trans>Disconnect</Trans>
-                      </WalletAction>
-                    </>
+                  {standaloneMode !== false && connectionType !== ConnectionType.GNOSIS_SAFE && (
+                    <WalletAction onClick={handleDisconnectClick}>
+                      <Trans>Disconnect</Trans>
+                    </WalletAction>
                   )}
                 </>
               )}