+++ apps/cowswap-frontend/src/modules/hooksStore/pure/AddCustomHookForm/styled.ts
@@ -5,6 +5,13 @@ export const Wrapper = styled.div`
   flex-direction: column;
   gap: 10px;
   padding: 10px;
+
+  > form {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    width: 100%;
+  }
 `
 
 export const Input = styled.div`