+++ libs/tokens/src/hooks/tokens/unsupported/useUnsupportedTokens.ts
@@ -1,7 +1,8 @@
 import { useAtomValue } from 'jotai'
 
 import { currentUnsupportedTokensAtom } from '../../../state/tokens/unsupportedTokensAtom'
+import { UnsupportedTokensState } from '../../../types';
 
-export function useUnsupportedTokens() {
-  return useAtomValue(currentUnsupportedTokensAtom)
+export function useUnsupportedTokens(): UnsupportedTokensState {
+  return useAtomValue(currentUnsupportedTokensAtom) || {}
 }