+++ libs/tokens/src/state/tokenLists/tokenListsStateAtom.ts
@@ -5,7 +5,6 @@ import { getJotaiMergerStorage } from '@cowprotocol/core'
 import { mapSupportedNetworks, SupportedChainId } from '@cowprotocol/cow-sdk'
 
 import {
-  ARBITRUM_ONE_TOKENS_LIST,
   DEFAULT_TOKENS_LISTS,
   GNOSIS_UNISWAP_TOKENS_LIST,
   LP_TOKEN_LISTS,
@@ -23,7 +22,8 @@ import { environmentAtom } from '../environmentAtom'
 const UNISWAP_TOKEN_LIST_URL: Record<SupportedChainId, string> = {
   [SupportedChainId.MAINNET]: UNISWAP_TOKENS_LIST,
   [SupportedChainId.GNOSIS_CHAIN]: GNOSIS_UNISWAP_TOKENS_LIST,
-  [SupportedChainId.ARBITRUM_ONE]: ARBITRUM_ONE_TOKENS_LIST,
+  [SupportedChainId.ARBITRUM_ONE]: UNISWAP_TOKENS_LIST,
+  [SupportedChainId.BASE]: UNISWAP_TOKENS_LIST,
   [SupportedChainId.SEPOLIA]: UNISWAP_TOKENS_LIST,
 }
 
@@ -46,14 +46,15 @@ export const userAddedListsSourcesAtom = atomWithStorage<ListsSourcesByNetwork>(
 export const allListsSourcesAtom = atom((get) => {
   const { chainId, useCuratedListOnly, isYieldEnabled } = get(environmentAtom)
   const userAddedTokenLists = get(userAddedListsSourcesAtom)
+  const userAddedTokenListsForChain = userAddedTokenLists[chainId] || []
 
   const lpLists = isYieldEnabled ? LP_TOKEN_LISTS : []
 
   if (useCuratedListOnly) {
-    return [get(curatedListSourceAtom), ...lpLists, ...userAddedTokenLists[chainId]]
+    return [get(curatedListSourceAtom), ...lpLists, ...userAddedTokenListsForChain]
   }
 
-  return [...DEFAULT_TOKENS_LISTS[chainId], ...lpLists, ...(userAddedTokenLists[chainId] || [])]
+  return [...(DEFAULT_TOKENS_LISTS[chainId] || []), ...lpLists, ...userAddedTokenListsForChain]
 })
 
 // Lists states
@@ -78,13 +79,14 @@ export const listsStatesMapAtom = atom((get) => {
   const allTokenListsInfo = get(listsStatesByChainAtom)
   const virtualListsState = get(virtualListsStateAtom)
   const userAddedTokenLists = get(userAddedListsSourcesAtom)
+  const useeAddedTokenListsForChain = userAddedTokenLists[chainId] || []
 
   const currentNetworkLists = {
     ...allTokenListsInfo[chainId],
     ...virtualListsState,
   }
 
-  const userAddedListSources = userAddedTokenLists[chainId].reduce<{ [key: string]: boolean }>((acc, list) => {
+  const userAddedListSources = useeAddedTokenListsForChain.reduce<{ [key: string]: boolean }>((acc, list) => {
     acc[list.source] = true
     return acc
   }, {})